refactor: create database directory
