+++ src/database/entity/Link.ts
